RadioGroupInput.test.tsx convert to RTL
